+++ clients/src/main/java/org/apache/kafka/common/network/SslChannelBuilder.java
@@ -196,6 +196,5 @@ public class SslChannelBuilder implements ChannelBuilder, ListenerReconfigurable
         public boolean complete() {
             return true;
         }
-
     }
 }
