+++ clients/src/test/java/org/apache/kafka/test/TestCondition.java
@@ -16,7 +16,6 @@
  */
 package org.apache.kafka.test;
 
-
 /**
  * Interface to wrap actions that are required to wait until a condition is met
  * for testing purposes.  Note that this is not intended to do any assertions.
